requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.3
channels==2.4.0
channels-redis==3.1.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.1.1
daphne==2.5.0
dj-database-url==0.5.0
Django==3.1.2
django-tinymce==3.2.0
django-webpack-loader==0.7.0
gunicorn==20.0.4
h11==0.11.0
hiredis==1.1.0
httptools==0.1.1
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
install==1.3.4
msgpack==1.0.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
unicodecsv==0.14.1
urllib3==1.25.10
uvicorn==0.12.2
uvloop==0.14.0
watchgod==0.6
websockets==8.1
zope.interface==5.1.2
